backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncmy==0.2.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
redis==5.0.4
celery==5.4.0
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.3
faker==25.3.0
aiofiles==23.2.1
tenacity==8.3.0
loguru==0.7.2
